requirements.txt
# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.3.0

# Data processing
pandas>=1.3.0

# Training utilities
tqdm>=4.62.0
tensorboardX>=2.4.0
scikit-learn>=0.24.0

# Configuration
pyyaml>=5.4.0
yacs>=0.1.8

# Model dependencies
timm>=0.4.12
einops>=0.4.0

# Transformers for BERT tokenizer
transformers>=4.20.0

# GIT model dependencies
azfuse @ git+https://github.com/microsoft/azfuse.git
pycocotools>=2.0.0
pycocoevalcap>=1.2

# Optional but recommended
boto3>=1.18.0
